execve: Add envp to the Linux notes about NULL pointers
During the review of static analysis results, we discovered a
functional, but non-portable, use of execve().  For example:

    char *cmd[] = { "/path/to/some/file", NULL };
    execve(cmd[0], cmd, NULL);

The call succeeds.  Yet, the static analysis tool (rightly)
pointed out that envp could be dereferenced.  But digging into
glibc and the kernel, it appears that like argv, envp when NULL
is treated as if it were an empty list.

So, to clear things up, I'm submitting this patch to update the
man page to indicate that envp is treated like argv.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>