Fix possible race condition in readlink.2 example
I noticed that the example in the readlink.2 man pages does error
checking for a race condition that would cause the value of the
symbolic link to get larger.  However, it doesn't handle the
opposite case, in which the value gets shorter.  (The NULL
terminator is always set at the old, longer offset.)  This could
cause the program to operate on uninitialized data.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>