perf_event_open.2: PERF_SAMPLE_BRANCH_STACK updates
This started out as just adding the new perf_event_open features
from Linux 3.11 (which was the addition of transactional memory
defines for PERF_SAMPLE_BRANCH_STACK samples) but turned into a
general cleanup of the PERF_SAMPLE_BRANCH_STACK documentation.

The main clarification is that at least one of the non-privilege values
must be set or else perf_event_open will return an EINVAL error.

Signed-off-by: Vince Weaver <vincent.weaver@maine.edu>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>