perf_event_open.2: Linux 3.12 rdpmc/mmap
It turns out that the perf_event mmap page rdpmc/time setting was
broken, dating back to the introduction of the feature.  Due
to a mistake with a bitfield, two different values mapped to
the same feature bit.

A new somewhat backwards compatible interface was introduced
in Linux 3.12.  A much longer report on the issue can be found
here:
   https://lwn.net/Articles/567894/

Signed-off-by: Vince Weaver <vincent.weaver@maine.edu>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>