fgetc.3, gets.3: Split gets(3) to isolate unsafe gets(3) to a page on its own
Currently man3/gets.3 documents various safe I/O functions, along
with the toxic "gets" function.

At the risk of being melodramatic, this strikes me as akin to
storing rat poison in a food cabinet, in the same style of
packaging as the food, but with a post-it note on it saying
"see warnings below".

I think such "never use this" functions should be quarantined
into their own manpages, rather than listing them alongside
sane functions.

The attached patch does this for "gets", moving the documentation
of the good functions from man3/gets.3 into man3/fgetc.3,
updating the SO links in the relevant functions to point at the
latter.

It then rewrites man3/gets.3 to spell out that "gets" is toxic
and should never be used (with a link to CWE-242 for good
measure).

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>