netdevice.7: Document SIOCGIFCONF case ifc_req==NULL
Add the missing description of the possibility to call SIOCGIFCONF
with ifc_req==NULL to determine the needed buffer size, as
described in
http://lkml.indiana.edu/hypermail/linux/kernel/0110.1/0506.html
and verified against source files net/core/dev_ioctl.c and
net/ipv4/devinet.c in the current kernel git tree.
This functionality has been present since the beginning of the 2.6
series. It's about time it gets documented.

While I'm at it, also generally clarify the section on
SIOCGIFCONF.

Signed-off-by: Tilman Schmidt <tilman@imap.cc>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>