perf_event_open.2 : PERF_SAMPLE_TRANSACTION support in Linux 3.13 (fwd)
The following patch adds descriptions of the new perf_event_open.2
PERF_SAMPLE_TRANSACTION sample type as added in Linux 3.13.

The descriptions are based on information provided by Andi Kleen,
both in the e-mail

    [PATCH 1/6] perf, core: Add generic transaction flags v5

sent to the linux-kernel list as well as an e-mail

    [PATCH] Document transaction flags in perf_event_open manpage

sent to the linux-man list.

The implementation is based heavily on the Intel Haswell
processor.  Documentation can be found at this page:
http://software.intel.com/en-us/blogs/2013/05/03/intelr-transactional-synchronization-extensions-intelr-tsx-profiling-with-linux-0
as well as in section 18.11.5.1 of volume 3 of the
Intel 64 and IA-32 Architecture Software Developer's Manual.

Also, someone with better manpage formatting skills than I have
should probably investigate why I can't get the last line of
the change to properly tab-align with the .I transaction
heading.

Cowritten-by: Andi Kleen <andi@firstfloor.org>
Signed-off-by: Vince Weaver <vincent.weaver@maine.edu>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>