close.2, getdomainname.2, getrlimit.2, madvise.2, sysinfo.2, umask.2, encrypt.3, fclose.3, fflush.3, lockf.3, rand.3, strtok.3, toupper.3, updwtmp.3, utmp.5, environ.7, hier.7, suffixes.7: Convert pages contain non-ASCII in source code comments to use UTF-8
Done using a slightly modified version of Peter Schiffer's
convert_to_utf_8.sh script. The script was modified so as *not*
a "coding:" marker to the groff source. For now, we'll only put
that marker on pages that contain non-ASCII characters in the
rendered text.

See https://bugzilla.kernel.org/show_bug.cgi?id=60807

Reported-by: Weizhou Pan <cs.wzpan@gmail.com>
Co-written-by: Peter Schiffer <pschiffe@redhat.com
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>