resolv.conf.5: DESCRIPTION: mention that the data is trusted
In a recent discussion about DNSSEC it was brought to my
attention that not all system administrators may understand
that the information in /etc/resolv.conf is fully trusted.
The resolver implementation in glibc treats /etc/resolv.conf
as a fully trusted source of DNS information and passes on
the AD-bit for DNSSEC as trusted.

This patch adds a clarifying sentence to make it absolutely
clear that indeed this source of information is trusted.

Signed-off-by: Carlos O'Donell <carlos@redhat.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>