sync.8: Page deleted
This page documents the sync command, which is provided by
coreutils. As such, the presence of this page in man-pages
makes no sense. And, anyway, the coreutils project has its
own sync(1) man page and "info" documentation. Furthermore,
the  man-pages sync(8) has crufty out-of-date and to some
extent irrelevant information.  So, now it's gone.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>