semget.2: Consolidate discussion on noninitialization of semaphores
The fact that semget() does not initialize the semaphores
in a new set was covered in two places (in DESCRIPTION
and BUGS). Consolidate these into one place (in NOTES)
and also point out that POSIX.1-2008 says that a future
version of the standard may require an implementation to
initialize the semaphores to 0.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>