link.2, open.2: Update AT_EMPTY_PATH and O_PATH documentation
The current text reflects the general worry in the kernel about
recipients of O_PATH fds being able to hardlink the referenced
files.  It turns out that it was possible to link these files
regardless of any possible security concerns.

Linux 3.11 removes the capability check in AT_EMPTY_PATH.
I expect that this functionality will be generally useful,
so let's document it better.

Signed-off-by: Andy Lutomirski <luto@amacapital.net>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>