stpcpy.3: Note some history of stpcpy()
Quoting Ingo:
I just noticed that the stpcpy(3) manual contains a speculation
that appears to be untrue on closer investigation:  That function
did not originate in MS DOS, but in Lattice C on AmigaDOS.

Here is a patch against the git master HEAD to fix that, and add
some more historical information.  To provide some background and
allow you to more easily verify the correctness of the patch, i'm
appending my mail to <misc@openbsd.org>, where i'm giving some
more details about the history and pointing to some primary
sources.  That mail also contains the (similar, but shorter)
patch i just committed to the OpenBSD manual page.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>