README: Explain usage of prefix when installing
Installing the man pages to /usr may not be the preferred option
when working on a system with a packaging system like apt or yum,
where /usr is reserved for packages and /usr/local or /opt is used
for user addons.

README should explain how to install to a path different to /usr.

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>