syslog.2: Rework text describing loglevels
The kernel header file mentioned in the discussion of the KERN_*
constants has morphed and is no longer exported inside glibc.
And the definitions of the constants themselves changed subtly
with kernel commit 04d2c8c83d0e3ac5f78aeede51babb3236200112.
So, rewrite the description of the constants to be a bit more
abstract.

Reported-by: Robert P. J. Day <rpjday@crashcourse.ca>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>