syscalls.2: Add prlimit64(2)
While trying to reconcile the new features in glibc with the
documented entries in the linux kernel man pages I noticed that
glibc exports prlimit64 for use by 32-bit applications (as does
the linux kernel), but that prlimit64 was not defined in the
syscalls list or in the prlimit-related page.

This is not the complete fix for this, but I don't have the time
to explain why and when prlimit64 should be used (or how it should
be used safely). Therefore I'm just patching the syscalls.2 list
to show that prlimit64 exists and was added in 2.6.36 (verified
with git by checking out the tags before and after).

Unless you've purposely excluded prlimit64 to avoid telling users
about it because it's complicated, please apply.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>