printf.3: Remove libc4 and libc5 details
Rich Felker noted that "scare text" in the man page warned about
the use of snprintf() on libc, and that some people had cited
this as a reason not to use snprintf().  Linux libc is now
ancient history, so there is no real need to keep that text.
But, while we're at it, we may as well clear out all of the
other ancient libc4 and libc5 pieces in the page. They are
nowadays more clutter than help.

Reported-by: Rich Felker <dalias@libc.org>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>