epoll_ctl.2: minor clarification concerning when a WAKEUP event is "processed"
I guess it should be obvious, but it never hurts to be explicit.
When the epoll file descriptor is closed all events are assumed
to be "processed".
Also if the event is removed with EPOLL_CTL_DEL or the EPOLLWAKEUP
flag is removed with EPOLL_CTL_MOD the event will be "processed".

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>