getgrent.3: Add ENOENT  and EAGAIN to error list
It's possible to get ENOENT returned from getgrent()
if the backend, for example say SSSD, isn't configured
or the daemon isn't running. The same can be said of any
of the NSS backend.

As POSIX does not list ENOENT, we can list it ourselves
and define it how we like.

I don't know how you handle errno values that are glibc
specific, but here is the patch that enhances getgrent(3)
to make users aware of what ENOENT is intended to mean
from glibc.

While I'm fixing one I might as well [add EAGAIN].

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>