mount.2: Clarify use of 'mountflags' and 'data' for MS_REMOUNT
Quoting Eric Biederman:

One thing that has come up recently (in 3 separate
implementations) is that mount(MS_REMOUNT|...,...) must include
all of the mount flags that need to be preserved.   People
creating read-only bind mounts tend to miss that and the locked
flags in mount namespaces.  That issue was flushed out now that
the kernel is now not allowing most mount flags to be cleared in
mount namespaces.  The interface is non-intuitive and we should
at least document the weirdness.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>