clock.3: The switch to the use of clock_gettime() was to improve *accuracy*
Quoting Vincent Lefevre:

    In glibc 2.17 and earlier, clock() was implemented on top of
    times(2).  For improved precision, since glibc 2.18, it is
                      ^^^^^^^^^^^^^^^^^^
    implemented on top of clock_gettime(2) (using the
    CLOCK_PROCESS_CPUTIME_ID clock).

This looks strange. The user doesn't seek improved precision, but
improved accuracy: if one gets more digits but the value itself is
less accurate (i.e. the error against the ideal value is larger),
this is bad. Perhaps changing "precision" to "accuracy" would be
correct (I assume that the real goal of the change was not just
improved precision, but more importantly the resulting improved
accuracy). I've reported a bug about the glibc documentation:

  https://sourceware.org/bugzilla/show_bug.cgi?id=17383

Reported-by: Vincent Lefevre <vincent@vinc17.net>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>