wait.2: ffix: add paragraph break before errno note in RETURN VALUE section
The text "Each of these calls sets errno to an appropriate
value in the case of an error." is not only for waitid.

This patch adds a paragraph break to move the errno note
to a new paragraph where it makes sense, as it applies to
all the wait* functions.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>