resolv.conf.5: Add missing no-tld-query option
Based on commit [1], ithe no-tld-query option exists for
resolv.conf configuration file.  Description of this option
is provided in [2]. This patch just copies this option
into resolv.conf.5 man page. Plus changes 'a' member
into 'an' before 'unqualified name as if it ...'
on the third line of [2]. Based on [3], this option
was added in glibc 2.14 as solving [4] bug.

[1] https://sourceware.org/git/?p=glibc.git;a=commitdiff;h=f87dfb1f11c01f2ccdc40d81e134cd06b32e28e8
[2] http://www.daemon-systems.org/man/resolv.conf.5.html man page.
[3] https://sourceware.org/git/?p=glibc.git;a=blob;f=NEWS;h=952f32af17e7fb49c4c1a305de673a13075bfaf5;hb=f87dfb1f11c01f2ccdc40d81e134cd06b32e28e8
[4] https://sourceware.org/bugzilla/show_bug.cgi?id=12734

Signed-off-by: Jan Chaloupka <jchaloup@redhat.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>