wordexp.3: Make it clear that WRDE_NOCMD prevents command substitution
The use of WRDE_NOCMD prevents command substitution. If the flag
WRDE_NOCMD is set then no command substitution shall occur and
the error WRDE_CMDSUB will be returned if such substitution is
requested when processing the words.

The manual page as-is makes it seem like the command substitution
occurs, and an error is returned *after* the substitution.
This patch clarifies that.

Signed-off-by: Carlos O'Donell <carlos@redhat.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>