proc.5: Document /proc/sys/vm/compact_memory
Since 2.6.36 [1] CONFIG_COMPACTION is supported in Linux (for a
detailed explanation see [2]). But in essence the contents of
/proc/buddyinfo shows how much contiguous areas of a certain
size are available. These numbers are not actually up to date
as the Linux kernel will implicitly trigger compaction
(+- defragmentation) when an allocation is done.  This makes it
difficult since you can't predict upfront whether or not a large
allocation will succeed or not.  By using
/proc/sys/vm/compact_memory you can explicitly trigger compaction
to occur and /proc/buddyinfo will give a more realistic view on
the available memory in contiguous areas.

This patch updates proc.5 to explain the usage of this file and is
inspired by Documentation/systcl/vm.txt [3]

[1] http://kernelnewbies.org/Linux_2_6_35#head-9cb0a1275559d40296da42efb7977896ac9edab7
[2] http://lwn.net/Articles/368869/
[3] http://lxr.free-electrons.com/source/Documentation/sysctl/vm.txt#L100

Signed-off-by: Elie De Brauwer <eliedebrauwer@gmail.com>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>