feature_test_macros.7: Update discussion of _FORTIFY_SOURCE
Since the initial implementation a lot more checks were added.
Describe all the checks would be too verbose (and would soon
fall out of date as more checks are added). So instead, describe
the kinds of checks that are done more generally.
Also a few other minor edits to the text.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>