tcp.7: Document removal of TCP_SYNQ_HSIZE
Looking over the man page for 'tcp' I came across a reference to
tuning the 'TCP_SYNQ_HSIZE' parameter when increasing
'tcp_max_syn_backlog' above 1024. However, this static sizing was
removed back in Linux 2.6.20 in favor of dynamic scaling - as
part of commit 72a3effaf633bcae9034b7e176bdbd78d64a71db.

I have included a patch below with reference to this commit, and
that the process detailed is not required on >= Linux 2.6.20.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>