qsort.3: alphasort() and versionsort() are not suitable for 'compar'
In glibc 2.10, the prototypes of alphasort() and versionsort()
were changed so that the arguments switched from 'const void *' to
'const struct dirent **', to match the POSIX.1-2008 specification
of alphasort(). As such, compiler warnings will result if
these functions are used as the arguments of qsort().

    warning: passing argument 4 of 'qsort' from incompatible
    pointer type
    expected '__compar_fn_t' but argument is of type
    'int (*)(const struct dirent **, const struct dirent **)'

Therefore, remove the ancient NOTES text suggesting that
alphasort() and versionsort() can be used as suitable
'compar' arguments for qsort().

Reported-by: Rob Landley <rob@landley.net>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>