CPU_SET.3: Clarify language about "available" cpus
The CPU_SET.3 man page uses the adjective "available" when
explaining what the argument to CPU_SET() means.  This is
confusing, since "available" isn't well-defined.  The kernel
has a set of adjectives (possible, present, online, and active)
that qualify cpus, but normally none of these are what the
cpu_set_t bit index means: it's just "which cpu", using the
kernel's internal numbering system, even if that cpu isn't
possible or present.

This change removes the word "available" and adds a sentence
warning that cpu sets may not be contiguous due to dynamic
cpu hotplug, etc.

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>