mmap.2: Remove text that implies that munmap() syncs MAP_SHARED mapping to file
The existing text in this page:

       MAP_SHARED Share  this  mapping.   Updates to the mapping
                  are visible to other processes that  map  this
                  file,  and are carried through to the underly‐
                  ing  file.   The  file  may  not  actually  be
                  updated until msync(2) or munmap() is called.

implies that munmap() will sync the mapping to the underlying
file. POSIX doesn't require this, and some light reading of the
code and some light testing (fsync() after munmap() of a large
file) also indicates that Linux doesn't do this.

See also this mail thread:

    Subject: munmap, msync: synchronization
    Newsgroups: gmane.linux.man
    Date: 2014-04-20 10:28:40 GMT
    http://thread.gmane.org/gmane.linux.man/5548

Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>