getopt.3: Remove crufty BUGS section
This piece of text in the man page is ancient (from
man-pages-1.20, circa 1998), and odd in the sense that it
describes a bug in POSIX that was (long ago) subsequently
fixed. As the standards committee noted, POSIX here seemed
to deviate from existing practice. The simplest fix is
to just remove this BUGS section.

See https://bugzilla.kernel.org/show_bug.cgi?id=90261

Reported-by: Guy Harris <guy@alum.mit.edu>
Signed-off-by: Michael Kerrisk <mtk.manpages@gmail.com>